requirements.txt
Scrapy
pymongo
Pillow
diffusers
transformers
diffusers
accelerate
datasets
torchvision
tensorboard